Changed to chain rule in gradient function
Basically reverted previous change and made the gradient function return the
dynamics object so can reuse adjoint calculation for different dprop lists
